Experiment Wiener search pruning
STATS_CHANGED

+++ aom_dsp/sum_squares.c
@@ -29,10 +29,10 @@ uint64_t aom_sum_squares_2d_i16_c(const int16_t *src, int src_stride, int width,
   return ss;
 }
 
-uint64_t aom_var_2d_u8_c(const uint8_t *src, int src_stride, int width,
-                         int height) {
+int64_t aom_var_2d_u8_c(const uint8_t *src, int src_stride, int width,
+                        int height) {
   int r, c;
-  uint64_t ss = 0, s = 0;
+  int64_t ss = 0, s = 0;
 
   for (r = 0; r < height; r++) {
     for (c = 0; c < width; c++) {
@@ -46,15 +46,15 @@ uint64_t aom_var_2d_u8_c(const uint8_t *src, int src_stride, int width,
   return (ss - s * s / (width * height));
 }
 
-uint64_t aom_var_2d_u16_c(const uint8_t *src, int src_stride, int width,
-                          int height) {
+int64_t aom_var_2d_u16_c(const uint8_t *src, int src_stride, int width,
+                         int height) {
   const uint16_t *srcp = CONVERT_TO_SHORTPTR(src);
   int r, c;
-  uint64_t ss = 0, s = 0;
+  int64_t ss = 0, s = 0;
 
   for (r = 0; r < height; r++) {
     for (c = 0; c < width; c++) {
-      const uint8_t v = srcp[c];
+      const uint16_t v = srcp[c];
       ss += v * v;
       s += v;
     }
